pyproject.toml
[project]
name = "rescue-ai"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "boto3==1.42.65",
  "fastapi==0.115.8",
  "uvicorn[standard]==0.30.6",
  "PyYAML==6.0.2",
]

[project.optional-dependencies]
dev = [
  "pytest==8.3.4",
  "pytest-cov==6.0.0",
  "httpx==0.27.2",
  "black==24.10.0",
  "flake8==7.1.1",
  "isort==5.13.2",
  "mypy==1.11.2",
  "types-PyYAML==6.0.12.20250915",
  "pylint==3.3.1",
]
inference = [
  "ultralytics==8.3.57",
]
postgres = [
  "psycopg[binary]==3.2.3",
]
batch = [
  "boto3==1.42.65",
  "psycopg[binary]==3.2.3",
]

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.12"
files = ["rescue_ai", "tests"]
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true

[tool.pylint.main]
py-version = "3.12"
